core: engine implement errorable interface
Therefore move engine creation outside of player, since engine now needs
to be added to the errorable list. This "list handling" has to remain
into core, and shouldn't be scattered everywhere in the code.

Notice that engine remains private to core, so that nobody deals with it
except for player.