libcaphe: fix dbus-invokator destruction
Objects must ensure uninhibit() on finalize. Trying to do that from the
calling code is slippery and create more problems.

Here is the best place to do it, the only important thing is to ensure
that we don't trigger any callback, because the objects usually passed
as data for the callbacks won't exist anymore.