fwk: use dim instead of grey in logs
A grey font is ok for a terminal with black background. However, on a white
background, it's not ok at all, it's almost unreadable. It should be dark
grey, but I can't set the color depending on the background.

So let' s just use 'dim' instead, and let the terminal decide what to do
with that.

On Terminator with black background, I still get the same
grey as before, so it's fine for me.

On Gnome terminal with white background, I get it all black. I would
prefer to have dark grey, but at least it's better than the normal grey
I had before.