feat: implement ui-related mpris2 properties and methods
This is a big commit !

Before, there was a clear distinction between core features (that had no
ui requirement) and ui features (that clearly depended on the ui).

Now we have this mpris2 feature that contain 'ifdef UI_ENABLED', and
therefore stands in between core and ui. And we can't have ui stuff
within the core, it's not supposed to be this way.

So I guess the best is just to move the feat out of core and ui, and let
them leave in a space of their own. The distinction between ui/core
features will be a bit more loose. It will be only visible in configure,
Makefile, and in the init code itself.